Actualizacion para seleccion aleatorio

+++ get_vpn.sh
@@ -5,7 +5,7 @@
 URLVPN=http://www.vpngate.net/api/iphone/
 
 #LOGS
-LISTSERVER=/tmp/vpngate.log
+LISTSERVER=/tmp/vpnservers.log
 
 #Data conexion seleccionada
 VPNSELECT=/tmp/openvpn.select.txt
@@ -13,14 +13,32 @@ VPNSELECT=/tmp/openvpn.select.txt
 #Configuracion OPENVPN
 VPNCFG=/tmp/openvpn.config
 
-#Recuperamos listado servidores disponibles
-wget -O ${LISTSERVER} ${URLVPN}
+AHORA=$(date +%s)
 
-#Contamos registros
-REG=$(wc -l ${LISTSERVER}| awk '{print $1}')
+#Recuperamos listado servidores disponibles
+function GetVPN ()
+{
+  #existe y esta dentro del timeout
+  if ! test -f ${LISTSERVER}
+  then
+    TFILEOUT=$(stat -c %Y ${LISTSERVER})
+    echo "${TFILEOUT} - ${AHORA}"
+    wget -O ${LISTSERVER} ${URLVPN}
+  fi
+}
 
 #Como esta codificado
 #HostName,IP,Score,Ping,Speed,CountryLong,CountryShort,NumVpnSessions,Uptime,TotalUsers,TotalTraffic,LogType,Operator,Message,OpenVPN_ConfigData_Base64
+
+########
+# PRINCIPAL
+########
+
+#Recuperamos listado servidores
+GetVPN
+
+#Contamos registros
+REG=$(wc -l ${LISTSERVER}| awk '{print $1}')
 
 #Creamos un Array de linea separado por ","
 DATA=$(head -2 ${LISTSERVER} | tail -1 )
@@ -33,13 +51,10 @@ IFS=${OIFS}
 
 echo "Total VPN Activas = ${REG}"
 
-
 #Podemos segun PAIS cualquier otro parametro, de momento a la mitad de registros
-#REGISTRO=$(expr ${REG} / 2)
-#echo "Registro Seleccionado = ${REGISTRO}"
 
-#Seleccion Segundo
-REGISTRO=$(date +%S)
+#Seleccion Registro aletorio
+REGISTRO=$(shuf -i 1-${REG} -n 1)
 echo "Registro Seleccionado = ${REGISTRO}"
 
 #Cojemos seleccion
@@ -73,16 +88,8 @@ echo "============================"
 #Extraemos Datos conexion y conectamos
 cat  ${VPNSELECT} | awk -F, '{print $15}' | base64 -d -i > ${VPNCFG}
 
+
 #Conectamos
 sudo openvpn --verb 0 --config ${VPNCFG}
 
-
-exit
-
---log file
---verb
-              0 -- No output except fatal errors.
-              1 to 4 -- Normal usage range.
-              5 -- Output R and W characters to the console for each packet read and write, uppercase is used for TCP/UDP packets and lowercase is used for TUN/TAP packets.
-              6 to 11 -- Debug info range (see errlevel.h for additional information on debug levels).
